Added makefile, separate object files
Header files do not generate data anymore.
