+++ src/Makefile
@@ -1,0 +1,31 @@
+#
+# Makefile automatically generated by MKMK V6.55
+# Tue Oct 27 15:25:00 2020
+#
+
+CC= cl65
+OBJS= desktop.o desktop-vectors.o desktop-menu.o desktop-icons.o
+CFLAGS= -t geos-cbm -O
+LDFLAGS= -t geos-cbm
+
+desktop.cvt: $(OBJS)
+	$(CC) $(LDFLAGS) -o $@ $(OBJS)
+
+desktop-res.h: desktop-res.grc
+
+desktop.o: desktop.c desktop.h desktop-res.h desktop-icons.h \
+           desktop-vectors.h
+
+desktop-vectors.o: desktop-vectors.c desktop-vectors.h
+
+desktop-menu.o: desktop-menu.c
+
+desktop-icons.o: desktop-icons.c desktop-icons.h
+
+desktop.d64: desktop.cvt
+	c1541 -format "desktop,sh" d64 $@
+	c1541 -attach $@ -write desktop.cvt $?
+
+
+clean:
+	$(RM) $(OBJS)
